fix: split oversized test files, fix lint errors
- Split 6 test files exceeding 500-line limit into 15 smaller files
- Remove unused asyncio and Callable imports from evaluate.py
- Fix Biome lint: remove unused 'success' import in pipeline.test.ts
- Fix Biome format: break long sortBy chain in validation.ts

Test file splits:
- test_surface_validation.py (1587) -> 4 files (analysis, rules, orchestrator, config)
- test_scenario_commands.py (1083) -> 3 files (list_create, evaluate, coverage_suggest)
- test_hints.py (891) -> 2 files (core, reports)
- test_adapters.py (639) -> 2 files (registry, generic)
- test_cli_wrapper.py (631) -> 2 files (types, execution)
- test_mcp_testing_server.py (576) -> 2 files (tools, calls)
