+++ packages/core/src/pipeline.test.ts
@@ -30,7 +30,6 @@ import {
 	resolveVariable,
 	resolveVariables,
 } from './pipeline.js';
-import { success } from './result.js';
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // TYPE GUARD TESTS
